Rewrite authentication mechanism
Almost the whole authentication mechanism is re-written.
Instead of writing and reading hashed credentials in local
storage, use the database to manage it.

A new model/table 'User' is added to the database, which stores
the credential of the user, e.g., username, hash (of password)
and salt. The modified authentication mechanism tries find whether
such user exists, and authenticate the user through comparing the
username and hashes.