Replace FilePicker with OS's file sharing mechanism
The images are now selected through OS's file sharing
mechanism. I.e., the user's preferred software (e.g., gallery)
will be opened for the user to select image, where the
content URI of the selected image is passed to this app for
encryption.

This way is much better than selecting files without a preview,
plus only the images can be selected.