Create a hash for image encryption/decryption
This hash is essentially a hash of (password + img_salt). It's
not stored anywhere in the internal storage or database, instead,
it's generated in runtime. This means that if one wants to retrieve
this key for decrypting the images, he must have known the password
already or else he will need to sneak into the program and somehow
retrieve the key in memory.