Impl AuthService using DI
This service class is injected to each activity. The functionalties
of authenticating a user and generating the image key in runtime
for image encyption and decryption are deleagted to this shared
service class. Such that, the activities are now manageing tasks
that are related to the view, and no more.