Fix permission bug for taking picture
A permission denial security exception is thrown when the user tries
to take a picture. The first time (on app startup) that the user tries
to take a picture, this functionality will work. However, the following
attempts will be rejected by the permission denial. This is caused by
incorrectly reusing the intent for ACTION_IMAGE_CAPTURE. This is now
fixed.